+++ apps/nextjs/src/app/aila/[id]/page.tsx
@@ -1,8 +1,5 @@
 "use client";
 
-import { useUser } from "@clerk/nextjs";
-import { redirect } from "next/navigation";
-
 import ChatPageContents from "../page-contents";
 
 interface ChatPageProps {
@@ -12,15 +9,10 @@ interface ChatPageProps {
 }
 
 export default function ChatPage({ params }: Readonly<ChatPageProps>) {
-  const user = useUser();
   const { id } = params;
   // For local development so that we can warm up the server
   if (id === "health") {
     return <>OK</>;
-  }
-
-  if (user.isLoaded && !user.isSignedIn) {
-    redirect(`/sign-in?next=/aila/${params.id}`);
   }
 
   return <ChatPageContents id={id} />;