+++ apps/nextjs/src/components/AppComponents/Chat/chat-left-hand-side.tsx
@@ -1,8 +1,6 @@
-import React from "react";
-
 import { Flex } from "@radix-ui/themes";
 
-import type { DemoContextProps } from "@/components/ContextProviders/Demo";
+import { useDemoLocking } from "@/hooks/useDemoLocking";
 
 import ChatLhsHeader from "./chat-lhs-header";
 import { ChatList } from "./chat-list";
@@ -11,20 +9,16 @@ import { ChatPanelArea } from "./chat-panel-area";
 import QuickActionButtons from "./chat-quick-buttons";
 
 type ChatLeftHandSideProps = {
-  isDemoLocked: boolean;
   showLessonMobile: boolean;
   setShowLessonMobile: (value: boolean) => void;
-  demo: DemoContextProps;
-  isDemoUser: boolean;
 };
 
 const ChatLeftHandSide = ({
-  isDemoLocked,
   showLessonMobile,
   setShowLessonMobile,
-  demo,
-  isDemoUser,
 }: Readonly<ChatLeftHandSideProps>) => {
+  const isDemoLocked = useDemoLocking();
+
   return (
     <Flex
       direction="column"
@@ -36,7 +30,6 @@ const ChatLeftHandSide = ({
       <ChatLhsHeader
         setShowLessonMobile={setShowLessonMobile}
         showLessonMobile={showLessonMobile}
-        isDemoUser={isDemoUser}
         showStreamingStatus={process.env.NEXT_PUBLIC_ENVIRONMENT !== "prd"}
       />
       <div>
@@ -44,7 +37,6 @@ const ChatLeftHandSide = ({
           <ChatList
             isDemoLocked={isDemoLocked}
             showLessonMobile={showLessonMobile}
-            demo={demo}
           />
         </ChatPanelArea>
         {!isDemoLocked && <QuickActionButtons />}