+++ apps/nextjs/src/stores/AilaStoresProvider.tsx
@@ -1,7 +1,9 @@
-import { useState, createContext, useContext, useEffect } from "react";
+import { useState, useRef, createContext, useContext, useEffect } from "react";
 
-import { useStore, type StoreApi } from "zustand";
+import invariant from "tiny-invariant";
+import { useStore, type StoreApi, type ExtractState } from "zustand";
 
+import { useLessonPlanTracking } from "@/lib/analytics/lessonPlanTrackingContext";
 import { type ChatStore, createChatStore } from "@/stores/chatStore";
 import {
   type ModerationStore,
@@ -11,56 +13,70 @@ import { trpc } from "@/utils/trpc";
 
 import { type LessonPlanStore, createLessonPlanStore } from "./lessonPlanStore";
 
-type AilaStoresContextProps = {
+export type AilaStores = {
   chat: StoreApi<ChatStore>;
   moderation: StoreApi<ModerationStore>;
   lessonPlan: StoreApi<LessonPlanStore>;
 };
+export type GetStore = <T extends keyof AilaStores>(
+  storeName: T,
+) => ExtractState<AilaStores[T]>;
 
-export const AilaStoresContext = createContext<
-  AilaStoresContextProps | undefined
->(undefined);
+export const AilaStoresContext = createContext<AilaStores | undefined>(
+  undefined,
+);
 
 export interface AilaStoresProviderProps {
   children: React.ReactNode;
   id: string;
 }
 
+export const buildStoreGetter =
+  (stores: Partial<AilaStores>) =>
+  <T extends keyof AilaStores>(storeName: T) => {
+    invariant(stores[storeName], `Store ${storeName} not initialised`);
+    return stores[storeName].getState() as ExtractState<AilaStores[T]>;
+  };
+
 export const AilaStoresProvider: React.FC<AilaStoresProviderProps> = ({
   children,
   id,
 }) => {
   const trpcUtils = trpc.useUtils();
+  const lessonPlanTracking = useLessonPlanTracking();
 
   const [stores] = useState(() => {
-    const moderationStore = createModerationStore({
+    const stores: Partial<AilaStores> = {};
+    const getStore = buildStoreGetter(stores);
+
+    stores.moderation = createModerationStore({
       id,
       trpcUtils,
+      getStore,
     });
-    const chatStore = createChatStore();
+    stores.chat = createChatStore(id, getStore, trpcUtils);
+    stores.lessonPlan = createLessonPlanStore({
+      id,
+      trpcUtils,
+      lessonPlanTracking,
+      getStore,
+    });
 
-    moderationStore.setState((state) => ({
-      ...state,
-      chatActions: chatStore.getState(),
-    }));
-
-    chatStore.setState((state) => ({
-      ...state,
-      moderationActions: moderationStore.getState(),
-    }));
-
-    return {
-      chat: chatStore,
-      moderation: moderationStore,
-      lessonPlan: createLessonPlanStore(id, trpcUtils),
-    };
+    return stores as AilaStores;
   });
 
   // Store initialisation
+  const haveInitialized = useRef(false);
   useEffect(() => {
+    // work around react strict mode double rendering
+    if (haveInitialized.current) {
+      return;
+    }
+    void stores.chat.getState().fetchInitialMessages();
     void stores.lessonPlan.getState().refetch();
     void stores.moderation.getState().fetchModerations();
-  }, [stores.lessonPlan, id, stores.moderation]);
+    haveInitialized.current = true;
+  }, [stores.lessonPlan, id, stores.moderation, stores.chat]);
 
   return (
     <AilaStoresContext.Provider value={stores}>