+++ apps/nextjs/src/stores/chatStore/index.ts
@@ -3,10 +3,13 @@ import { aiLogger } from "@oakai/logger";
 import type { ChatRequestOptions, CreateMessage } from "ai";
 import { createStore } from "zustand";
 
-import type { ModerationStore } from "../moderationStore";
+import type { TrpcUtils } from "@/utils/trpc";
+
+import type { GetStore } from "../AilaStoresProvider";
 import { logStoreUpdates } from "../zustandHelpers";
 import { handleAppend } from "./stateActionFunctions/handleAppend";
 import { handleExecuteQueuedAction } from "./stateActionFunctions/handleExecuteQueuedAction";
+import { handleFetchInitialMessages } from "./stateActionFunctions/handleFetchInitialMessages";
 import { handleScrollToBottom } from "./stateActionFunctions/handleScrollToBottom";
 import { handleSetMessages } from "./stateActionFunctions/handleSetMessages";
 import { handleStop } from "./stateActionFunctions/handleStop";
@@ -18,7 +21,7 @@ const log = aiLogger("chat:store");
 
 export type AiSdkActions = {
   stop: () => void;
-  reload: () => void;
+  reload: () => Promise<string | null | undefined>;
   append: (
     message: AiMessage | CreateMessage,
     chatRequestOptions?: ChatRequestOptions | undefined,
@@ -35,9 +38,10 @@ export type AilaStreamingStatus =
   | "Idle";
 
 export type ChatStore = {
-  moderationActions?: Pick<ModerationStore, "fetchModerations">;
+  id: string;
   ailaStreamingStatus: AilaStreamingStatus;
 
+  initialMessages: AiMessage[];
   stableMessages: ParsedMessage[];
   streamingMessage: ParsedMessage | null;
   queuedUserAction: string | null;
@@ -53,6 +57,7 @@ export type ChatStore = {
   setAiSdkActions: (actions: AiSdkActions) => void;
   setMessages: (messages: AiMessage[], isLoading: boolean) => void;
   setInput: (input: string) => void;
+  setChatAreaRef: (ref: React.RefObject<HTMLDivElement>) => void;
 
   // Action functions
   executeQueuedAction: () => void;
@@ -60,12 +65,20 @@ export type ChatStore = {
   stop: () => void;
   streamingFinished: () => void;
   scrollToBottom: () => void;
+  fetchInitialMessages: () => Promise<void>;
 };
 
-export const createChatStore = (initialValues: Partial<ChatStore> = {}) => {
+export const createChatStore = (
+  id: string,
+  getStore: GetStore,
+  trpcUtils: TrpcUtils,
+  initialValues: Partial<ChatStore> = {},
+) => {
   const chatStore = createStore<ChatStore>((set, get) => ({
+    id,
     moderationActions: undefined, // Passed in the provider
     ailaStreamingStatus: "Idle",
+    initialMessages: [],
     stableMessages: [],
     streamingMessage: null,
     queuedUserAction: null,
@@ -76,8 +89,8 @@ export const createChatStore = (initialValues: Partial<ChatStore> = {}) => {
     // From AI SDK
     aiSdkActions: {
       stop: () => {},
-      reload: () => {},
-      append: async () => Promise.resolve(""),
+      reload: () => Promise.resolve(null),
+      append: () => Promise.resolve(""),
     },
 
     // Setters
@@ -90,9 +103,10 @@ export const createChatStore = (initialValues: Partial<ChatStore> = {}) => {
     executeQueuedAction: handleExecuteQueuedAction(set, get),
     append: handleAppend(set, get),
     stop: handleStop(set, get),
-    setMessages: handleSetMessages(set, get),
+    setMessages: handleSetMessages(getStore, set, get),
     streamingFinished: handleStreamingFinished(set, get),
     scrollToBottom: handleScrollToBottom(set, get),
+    fetchInitialMessages: handleFetchInitialMessages(set, get, trpcUtils),
 
     ...initialValues,
   }));