+++ apps/nextjs/src/stores/chatStore/stateActionFunctions/handleSetMessages.ts
@@ -1,13 +1,16 @@
 import invariant from "tiny-invariant";
 
-import type { AilaStreamingStatus, ChatStore } from "..";
+import type { GetStore } from "@/stores/AilaStoresProvider";
+
+import type { AilaStreamingStatus } from "..";
 import { calculateStreamingStatus } from "../actions/calculateStreamingStatus";
 import { getNextStableMessages, parseStreamingMessage } from "../parsing";
-import type { AiMessage } from "../types";
+import type { ChatSetter, ChatGetter, AiMessage } from "../types";
 
 export function handleSetMessages(
-  set: (partial: Partial<ChatStore>) => void,
-  get: () => ChatStore,
+  getStore: GetStore,
+  set: ChatSetter,
+  get: ChatGetter,
 ) {
   function handleChangedAilaStreamingStatus(
     ailaStreamingStatus: AilaStreamingStatus,
@@ -16,16 +19,20 @@ export function handleSetMessages(
       void get().executeQueuedAction();
     }
     if (ailaStreamingStatus === "Idle") {
-      const { moderationActions } = get();
-      invariant(moderationActions, "Passed into store in provider");
-      void moderationActions.fetchModerations();
+      void getStore("moderation").fetchModerations();
     }
+  }
+
+  function lastMessageIsUser(messages: AiMessage[]) {
+    return messages[messages.length - 1]?.role === "user";
   }
 
   return (messages: AiMessage[], isLoading: boolean) => {
     const originalStreamingStatus = get().ailaStreamingStatus;
 
     if (!isLoading) {
+      // The AI SDK isn't loading: we're idle and all messages are stable
+
       const nextStableMessages = getNextStableMessages(
         messages,
         get().stableMessages,
@@ -38,7 +45,23 @@ export function handleSetMessages(
         streamingMessage: null,
         ailaStreamingStatus: "Idle",
       });
+    } else if (lastMessageIsUser(messages)) {
+      // AI SDK is loading without a message from the API: we're waiting for a response
+
+      const nextStableMessages = getNextStableMessages(
+        messages,
+        get().stableMessages,
+      );
+      set({
+        ...(nextStableMessages && {
+          stableMessages: nextStableMessages,
+        }),
+        streamingMessage: null,
+        ailaStreamingStatus: "RequestMade",
+      });
     } else {
+      // AI SDK is loading with a message from the API: we're streaming
+
       const currentMessageData = messages[messages.length - 1];
       invariant(currentMessageData, "Should have at least one message");
       const streamingMessage = parseStreamingMessage(