+++ apps/nextjs/src/stores/chatStore/types.ts
@@ -1,11 +1,16 @@
 import type { MessagePart } from "@oakai/aila/src/protocol/jsonPatchProtocol";
 import type { Message as AiMessage } from "ai";
+import type { StoreApi } from "zustand";
+
+import type { ChatStore } from ".";
 
 export { AiMessage };
 
-// NOTE: The shape of this type is still in flux
 export type ParsedMessage = AiMessage & {
   parts: MessagePart[];
   hasError: boolean;
   isEditing: boolean;
 };
+
+export type ChatSetter = StoreApi<ChatStore>["setState"];
+export type ChatGetter = StoreApi<ChatStore>["getState"];