+++ apps/nextjs/src/stores/moderationStore/actionFunctions/handleToxicModeration.tsx
@@ -1,17 +1,18 @@
 import type { PersistedModerationBase } from "@oakai/core/src/utils/ailaModeration/moderationSchema";
-import invariant from "tiny-invariant";
+
+import type { GetStore } from "@/stores/AilaStoresProvider";
 
 import type { ModerationStore } from "..";
 
-export const handleToxicModeration = (
-  mod: PersistedModerationBase | null,
-  set: (state: Pick<ModerationStore, "toxicModeration">) => void,
-  get: () => ModerationStore,
-) => {
-  set({ toxicModeration: mod });
-  const { chatActions } = get();
-  invariant(chatActions, "Passed into store in provider");
-  chatActions.setMessages([], false);
-
-  //  @TODO setOverrideLessonPlan({});
-};
+export const handleToxicModeration =
+  (
+    getStore: GetStore,
+    set: (state: Pick<ModerationStore, "toxicModeration">) => void,
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    get: () => ModerationStore,
+  ) =>
+  (mod: PersistedModerationBase | null) => {
+    set({ toxicModeration: mod });
+    getStore("chat").setMessages([], false);
+    getStore("lessonPlan").resetStore();
+  };