+++ apps/nextjs/src/stores/moderationStore/index.ts
@@ -2,17 +2,15 @@ import type { PersistedModerationBase } from "@oakai/core/src/utils/ailaModerati
 import type { Moderation } from "@prisma/client";
 import { create } from "zustand";
 
+import type { GetStore } from "@/stores/AilaStoresProvider";
 import type { TrpcUtils } from "@/utils/trpc";
 
-import type { ChatStore } from "../chatStore";
 import { logStoreUpdates } from "../zustandHelpers";
 import { handleFetchModerations } from "./actionFunctions/handleFetchModeration";
 import { handleToxicModeration } from "./actionFunctions/handleToxicModeration";
 import { handleUpdateModerationState } from "./actionFunctions/handleUpdateModerationState";
 
 export type ModerationStore = {
-  chatActions?: Pick<ChatStore, "setMessages">;
-
   id: string;
   moderations: Moderation[] | [];
   isModerationsLoading: boolean | null;
@@ -32,15 +30,16 @@ export type ModerationStore = {
 export const createModerationStore = ({
   id,
   initialValues = {},
+  getStore,
   trpcUtils,
 }: {
   id: string;
   initialValues?: Partial<ModerationStore>;
+  getStore: GetStore;
   trpcUtils: TrpcUtils;
 }) => {
   const moderationStore = create<ModerationStore>((set, get) => ({
     id,
-    chatActions: undefined, // Passed in the provider
     moderations: [],
     isModerationsLoading: null,
     toxicInitialModeration: null,
@@ -51,12 +50,8 @@ export const createModerationStore = ({
     setIsModerationsLoading: (isModerationsLoading) =>
       set({ isModerationsLoading }),
 
-    updateToxicModeration: (mod) => {
-      handleToxicModeration(mod, set, get);
-    },
-    updateModerationState: (mod) => {
-      handleUpdateModerationState(mod, set, get);
-    },
+    updateToxicModeration: handleToxicModeration(getStore, set, get),
+    updateModerationState: handleUpdateModerationState(set, get),
 
     fetchModerations: handleFetchModerations(set, get, trpcUtils),