+++ CHANGE_LOG.md
@@ -1,3 +1,21 @@
+## [1.25.5](https://github.com/oaknational/oak-ai-lesson-assistant/compare/v1.25.4...v1.25.5) (2025-02-24)
+
+
+### Bug Fixes
+
+* getChat overfetching - move initialMessages to store ([#550](https://github.com/oaknational/oak-ai-lesson-assistant/issues/550)) ([0ea1694](https://github.com/oaknational/oak-ai-lesson-assistant/commit/0ea169487451e97ab5975c7db3bb7fb511c223a8))
+
+## [1.25.4](https://github.com/oaknational/oak-ai-lesson-assistant/compare/v1.25.3...v1.25.4) (2025-02-20)
+
+
+### Bug Fixes
+
+* maths lesson plan overrides ([#557](https://github.com/oaknational/oak-ai-lesson-assistant/issues/557)) ([e84f44c](https://github.com/oaknational/oak-ai-lesson-assistant/commit/e84f44c7cf76397d35ed46a3c7d0310cbd2457e5))
+
+## [1.25.3](https://github.com/oaknational/oak-ai-lesson-assistant/compare/v1.25.2...v1.25.3) (2025-02-19)
+
+## [1.25.2](https://github.com/oaknational/oak-ai-lesson-assistant/compare/v1.25.1...v1.25.2) (2025-02-17)
+
 ## [1.25.1](https://github.com/oaknational/oak-ai-lesson-assistant/compare/v1.25.0...v1.25.1) (2025-02-14)